#307 - Resolve dir path naming conflicts
When saving another trufos object which has a
similar title than a previously saved trufos
object, the dir name might turn out to be the same
for both objects, causing one object to overwrite
the other.

This commit solves this issue by appending a
number to the dir name if the dir name is already
taken. The number is incremented until a unique
dir name is found.