#307 - Resolve dir path naming conflicts (#362)
* #307 - Resolve dir path naming conflicts

When saving another trufos object which has a
similar title than a previously saved trufos
object, the dir name might turn out to be the same
for both objects, causing one object to overwrite
the other.

This commit solves this issue by appending a
number to the dir name if the dir name is already
taken. The number is incremented until a unique
dir name is found.

* #307 - Remove unnecessary dir path taken check

This is not necessary as the
loop above only breaks when
the condition is already met.

* #307 - Fix formatting

---------

Co-authored-by: Kevin Werber <kevin.werber@exxeta.com>