145 - Add HTTP header suggestion feature using shadcn's Command compo… (#416)
* #145 - Add HTTP header suggestion feature using shadcn's Command component
I add a new feature for HTTP header suggestions using shadcn's Command component to make the UI more user-friendly.
I create a constants folder with an index.ts file to store common headers and their values in one place.
Everything is connected in the HeaderTab component, where users can see and select header suggestions easily.
This helps keep the code clean and makes it easier to update the headers in the future.

* #416 - Update: created a new HeaderRow component to improve performance.
Earlier, we were filtering header key and value suggestions on every render inside
the HeaderTab component, which was not optimized.
To fix this, I moved the filtering logic into a separate HeaderRow component and
used the useMemo hook so that filtering only happens when the header key or value changes.
Then I called HeaderRow from HeaderTab and passed the required props to it like
header, index, and the update and delete functions.

* #416 - use popover to show header autocompletion

* #416 - fix flickering

* #416 - enable new headers by default

* #416 - prettier

---------

Co-authored-by: SoulKa <louiswilke@gmx.de>