feat(545): Add close Collection Button and closing Logic (#570)
* 545-Add close collection button and closing logic
Adding the closing button with the given close icon.
Use same functionallity as not found collections to close a collection
To ensure proper updates the collections state is filtered.

With these changes, collections can be removed from the collections dropdown menu without deleting them from the system.

* 545- Adapt button style to fit the overall style
Changing the text style, adding a hover functionality.
Fixing the centering of text by changing the span to p.

* 545-Fixing alignment to the right side of the dropdown menu
The close button is now aligned to the right side of the dropdown menu to fit the overall style.

* 545-Removing closing button for first collection
The closing button is removed for the first collection as this collection is required for proper execution.